+++ game/src/main_menu/mainmenu.rs
@@ -5,11 +5,12 @@ use crate::backend::{
 };
 use crate::main_menu::mainmenu::Message::{Exit, NewGame, Start};
 use crate::RLResult;
-use ggez::event::MouseButton;
+use crate::main_menu::button::Button;
 use ggez::graphics::Color;
 use ggez::mint::Point2;
 use ggez::{graphics, Context, GameResult};
 use std::sync::mpsc::{channel, Receiver, Sender};
+use ggez::event::MouseButton;
 
 #[derive(Copy, Clone, Debug, PartialEq)]
 pub enum Message {
@@ -19,39 +20,13 @@ pub enum Message {
 }
 
 #[derive(Debug)]
-struct Button<T: Clone> {
-    text: String,
-    img: Option<graphics::Image>,
-    message: Message,
-    sender: Sender<T>,
-    rect: graphics::Rect,
-    color: Color,
-}
-
-impl Button<Message> {
-    fn pressed(&self) {
-        dbg!("Pressed {:?}", self.message);
-    }
-
-    fn is_clicked(&self, mouse_pos: Point2<f32>) -> bool {
-        self.rect.contains(mouse_pos)
-    }
-    fn click(&mut self, mouse_pos: Point2<f32>) {
-        if self.is_clicked(mouse_pos) {
-            self.pressed();
-            self.sender.send(self.message).unwrap();
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct MainMenu<Message: Clone> {
-    buttons: Vec<Button<Message>>,
+pub struct MainMenu {
+    buttons: Vec<Button>,
     receiver: Receiver<Message>,
     sender: Sender<Message>,
 }
 
-fn draw_button(ctx: &mut Context, btn: &Button<Message>) -> GameResult<graphics::Mesh> {
+fn draw_button(ctx: &mut Context, btn: &Button) -> GameResult<graphics::Mesh> {
     let mb = &mut graphics::MeshBuilder::new();
 
     mb.rectangle(graphics::DrawMode::fill(), btn.rect, btn.color)?;
@@ -59,7 +34,7 @@ fn draw_button(ctx: &mut Context, btn: &Button<Message>) -> GameResult<graphics:
     Ok(graphics::Mesh::from_data(ctx, mb.build()))
 }
 
-impl<Message: Clone> Default for MainMenu<Message> {
+impl Default for MainMenu {
     fn default() -> Self {
         let (sender, receiver) = channel();
 
@@ -96,7 +71,7 @@ impl<Message: Clone> Default for MainMenu<Message> {
     }
 }
 
-impl Screen for MainMenu<Message> {
+impl Screen for MainMenu {
     fn update(&mut self, ctx: &mut Context) -> RLResult<StackCommand> {
         //handle buttons
         if ctx.mouse.button_pressed(MouseButton::Left) {
